+++ src/dataset-builder/CohortEditor.ts
@@ -11,7 +11,6 @@ import {
   AnyCriteria,
   Cohort,
   CriteriaGroup,
-  DatasetParticipantCountResponse,
   formatCount,
   ProgramDataListCriteria,
   ProgramDataRangeCriteria,
@@ -19,6 +18,7 @@ import {
 import {
   DataRepo,
   DatasetModel,
+  DatasetParticipantCountResponse,
   SnapshotBuilderDomainOption,
   SnapshotBuilderOption,
   SnapshotBuilderProgramDataListItem,
